+++ examples/.gitignore
@@ -1,0 +1,4 @@
+*.d.ts
+*.js
+node_modules
+dist/
