--- packages/.gitignore
@@ -1,1 +1,0 @@
-!/projen-common.js
