+++ packages/cdk8s-cli/.gitignore
@@ -35,6 +35,9 @@ jspm_packages/
 .yarn-integrity
 # parcel-bundler cache (https://parceljs.org/)
 .cache
+# jest-junit artifacts
+/test-reports/
+junit.xml
 /coverage
 /lib
 /dist
