+++ packages/cdk8s-cli/package.json
@@ -31,21 +31,22 @@
     "organization": true
   },
   "devDependencies": {
-    "@types/fs-extra": "^8.1.0",
-    "@types/jest": "^26.0.7",
-    "@types/json-schema": "^7.0.5",
+    "@types/fs-extra": "^9.0.4",
+    "@types/jest": "^26.0.15",
+    "@types/json-schema": "^7.0.6",
     "@types/node": "^10.17.0",
     "@typescript-eslint/eslint-plugin": "^4.3.0",
     "@typescript-eslint/parser": "^4.3.0",
-    "eslint": "^6.8.0",
-    "eslint-import-resolver-node": "^0.3.3",
-    "eslint-import-resolver-typescript": "^2.0.0",
-    "eslint-plugin-import": "^2.20.2",
-    "jest": "^26.4.2",
+    "eslint": "^7.14.0",
+    "eslint-import-resolver-node": "^0.3.4",
+    "eslint-import-resolver-typescript": "^2.3.0",
+    "eslint-plugin-import": "^2.22.1",
+    "jest": "^26.6.3",
+    "jest-junit": "^12",
     "json-schema": "^0.2.5",
-    "projen": "^0.3.168",
+    "projen": "^0.5.0",
     "standard-version": "^9.0.0",
-    "ts-jest": "^26.1.0",
+    "ts-jest": "^26.4.4",
     "typescript": "^3.9.5"
   },
   "peerDependencies": {},
@@ -54,13 +55,13 @@
     "cdk8s": "0.0.0",
     "codemaker": "^1.14.1",
     "constructs": "^3.2.34",
-    "fs-extra": "^8.1.0",
+    "fs-extra": "^9.0.1",
     "jsii-pacmak": "^1.14.1",
-    "jsii-srcmak": "^0.1.36",
-    "json2jsii": "^0.1.104",
+    "jsii-srcmak": "^0.1.115",
+    "json2jsii": "^0.1.113",
     "sscaff": "^1.2.0",
-    "yaml": "^1.7.2",
-    "yargs": "^15.1.0"
+    "yaml": "^1.10.0",
+    "yargs": "*"
   },
   "bundledDependencies": [],
   "keywords": [
@@ -69,6 +70,9 @@
     "k8s",
     "kubernetes"
   ],
+  "engines": {
+    "node": ">= 10.17.0"
+  },
   "main": "lib/index.js",
   "license": "Apache-2.0",
   "start": {
@@ -148,6 +152,15 @@
     "testMatch": [
       "**/__tests__/**/*.ts?(x)",
       "**/?(*.)+(spec|test).ts?(x)"
+    ],
+    "reporters": [
+      "default",
+      [
+        "jest-junit",
+        {
+          "outputDirectory": "test-reports"
+        }
+      ]
     ],
     "preset": "ts-jest",
     "globals": {
