+++ packages/cdk8s-plus-17/package.json
@@ -32,26 +32,27 @@
     "organization": true
   },
   "devDependencies": {
-    "@types/jest": "^26.0.7",
+    "@types/jest": "^26.0.15",
     "@types/minimatch": "^3.0.3",
     "@types/node": "^10.17.0",
     "@typescript-eslint/eslint-plugin": "^4.3.0",
     "@typescript-eslint/parser": "^4.3.0",
     "cdk8s": "^0.0.0",
     "constructs": "3.2.34",
-    "eslint": "^6.8.0",
-    "eslint-import-resolver-node": "^0.3.3",
-    "eslint-import-resolver-typescript": "^2.0.0",
-    "eslint-plugin-import": "^2.20.2",
-    "jest": "^26.4.2",
+    "eslint": "^7.14.0",
+    "eslint-import-resolver-node": "^0.3.4",
+    "eslint-import-resolver-typescript": "^2.3.0",
+    "eslint-plugin-import": "^2.22.1",
+    "jest": "^26.6.3",
+    "jest-junit": "^12",
     "jsii": "^1.14.1",
     "jsii-diff": "^1.14.1",
-    "jsii-docgen": "^1.3.2",
+    "jsii-docgen": "^1.4.96",
     "jsii-pacmak": "^1.14.1",
     "json-schema": "^0.2.5",
-    "projen": "^0.3.168",
+    "projen": "^0.5.0",
     "standard-version": "^9.0.0",
-    "ts-jest": "^26.1.0",
+    "ts-jest": "^26.4.4",
     "typescript": "^3.9.5"
   },
   "peerDependencies": {
@@ -160,6 +161,15 @@
     "testMatch": [
       "**/__tests__/**/*.ts?(x)",
       "**/?(*.)+(spec|test).ts?(x)"
+    ],
+    "reporters": [
+      "default",
+      [
+        "jest-junit",
+        {
+          "outputDirectory": "test-reports"
+        }
+      ]
     ],
     "preset": "ts-jest",
     "globals": {
